# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agri-PoD Inventory & Distribution System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List vehicles
      responses:
        "200":
          description: Vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Register vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/drivers:
    get:
      operationId: listDrivers
      tags: [vehicles]
      summary: List drivers
      responses:
        "200":
          description: Drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [vehicles]
      summary: Register driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

  # DISPATCH
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/activate:
    post:
      operationId: activateUser
      tags: [users]
      summary: Activate user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/deactivate:
    post:
      operationId: deactivateUser
      tags: [users]
      summary: Deactivate user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        vehicleCode:
          type: string
        plateNumber:
          type: string
        vehicleType:
          type: string
        capacity:
          type: number
          nullable: true
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        gpsDeviceId:
          type: string
          nullable: true
        status:
          type: string
          enum: [Active, Inactive, InTransit, Maintenance]
        lastLatitude:
          type: number
          nullable: true
        lastLongitude:
          type: number
          nullable: true
        lastPing:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, vehicleCode, plateNumber, vehicleType, status, createdAt]
    CreateVehicleBody:
      type: object
      properties:
        plateNumber:
          type: string
        vehicleType:
          type: string
        capacity:
          type: number
          nullable: true
        driverId:
          type: integer
          nullable: true
        gpsDeviceId:
          type: string
          nullable: true
      required: [plateNumber, vehicleType]
    Driver:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        licenseNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
        vehicleId:
          type: integer
          nullable: true
        vehiclePlate:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, isActive, createdAt]
    CreateDriverBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        licenseNumber:
          type: string
          nullable: true
      required: [fullName]

    # DISPATCH
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
          enum: [SystemAdmin, ProjectManager, ProcurementOfficer, WarehouseManager, DistrictCoordinator, FieldOfficer, Driver, MEOfficer, Auditor]
        districtId:
          type: integer
          nullable: true
        districtName:
          type: string
          nullable: true
        isActive:
          type: boolean
        lastLogin:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, username, fullName, role, isActive, createdAt]
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
          enum: [SystemAdmin, ProjectManager, ProcurementOfficer, WarehouseManager, DistrictCoordinator, FieldOfficer, Driver, MEOfficer, Auditor]
        password:
          type: string
        districtId:
          type: integer
          nullable: true
      required: [username, fullName, role, password]
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
          enum: [SystemAdmin, ProjectManager, ProcurementOfficer, WarehouseManager, DistrictCoordinator, FieldOfficer, Driver, MEOfficer, Auditor]
        districtId:
          type: integer
          nullable: true